fix: setup.py can now find numpy
The old setup.py code was unable to build the fastgrab module, due to missing
numpy during the setup stage. It's unlikely that anyone ever installed this
library via PyPi in the past, due to this oversight.

However, it's possible to patch the build process class to do late injection
of the numpy headers path. This technique comes from the following post:

https://stackoverflow.com/a/21621689/8874388

It works by telling setuptools that we need numpy, and then modifying the
build process class to add numpy's path before we actually build fastgrab.

Further fixes were performed to make it buildable on Python 3.12 with the
latest libraries.